All works at Bedwyn

+++ build.sh
@@ -3,18 +3,22 @@ echo Building web chat project Simon Hewitt 806068
 
 rm -rf *.class
 echo stopping webPublisher
-kill -9  `lsof -ti:8080`
+if [[ -n $(lsof -ti:8080) ]]; then
+  kill -9  `lsof -ti:8080`
+fi
 jobs
 
 
 echo building Publisher
-javac WebChatPublisher.java
+export CLASSPATH=${CLASSPATH}:.
+#echo $CLASSPATH
+javac -cp . WebChatPublisher.java
 echo building server
-javac chatserver/*.java
+javac -cp webchatclient/ chatserver/*.java
 echo Running publisher
 java WebChatPublisher&
 echo Wait for server to start up
-sleep 4
+sleep 2
 
 cd webchatclient
 echo Generate WSDL classes etc
